no more `IndexedDbMigration` but all inside `IndexedDbDatabase`
