+++ packages/platform-browser/examples/indexedDb.ts
@@ -1,10 +1,4 @@
-import {
-  IndexedDb,
-  IndexedDbDatabase,
-  IndexedDbMigration,
-  IndexedDbTable,
-  IndexedDbVersion
-} from "@effect/platform-browser"
+import { IndexedDb, IndexedDbDatabase, IndexedDbTable, IndexedDbVersion } from "@effect/platform-browser"
 import { Effect, Layer, Schema } from "effect"
 
 class UserTable1 extends IndexedDbTable.make(
@@ -37,9 +31,7 @@ class UserTable2 extends IndexedDbTable.make(
 class DatabaseVersion1 extends IndexedDbVersion.make(UserTable1) {}
 class DatabaseVersion2 extends IndexedDbVersion.make(UserTable2) {}
 
-const CurrentDatabaseVersion = DatabaseVersion2
-
-class Migration extends IndexedDbMigration.make(DatabaseVersion1, (api) =>
+class Migration extends IndexedDbDatabase.make(DatabaseVersion1, (api) =>
   Effect.gen(function*() {
     yield* api.createObjectStore("user")
     yield* api.createIndex("user", "email")
@@ -54,14 +46,14 @@ class Migration extends IndexedDbMigration.make(DatabaseVersion1, (api) =>
     }))
 {}
 
-const MainLayer = IndexedDbDatabase.layer("database", Migration).pipe(
+const MainLayer = Migration.layer("database").pipe(
   Layer.provide(IndexedDb.layerWindow)
 )
 
 export class IndexedDbService extends Effect.Service<IndexedDbService>()("IndexedDbService", {
   dependencies: [MainLayer],
   effect: Effect.gen(function*() {
-    const api = yield* CurrentDatabaseVersion.api
+    const api = yield* Migration.getQueryBuilder
     return api
   })
 }) {}
